moviedb-374 moviedb Move movieinteger.py into global constants.py
